// src/types/clinicalHistory.ts
export interface AccidentEntry {
  fecha: string;
  nombre_empresa: string;
  tipo_accidente: string;
  parte_cuerpo_afectada: string;
  dias_ausencia_laboral: string;
  secuelas: string;
}